Move tests to chat package
