+++ sdk/communication/azure-communication-chat/src/test/java/com/azure/android/communication/chat/IdentifierSerialzationExceptionTests.java
@@ -1,7 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.communication.common;
+package com.azure.android.communication.chat;
+
+import com.azure.android.communication.chat.models.CommunicationCloudEnvironmentModel;
+import com.azure.android.communication.chat.models.CommunicationIdentifierModel;
+import com.azure.android.communication.chat.models.CommunicationUserIdentifierModel;
+import com.azure.android.communication.chat.models.MicrosoftTeamsUserIdentifierModel;
+import com.azure.android.communication.chat.models.PhoneNumberIdentifierModel;
 
 import org.junit.Rule;
 import org.junit.Test;
@@ -12,7 +18,7 @@ import org.junit.runners.Parameterized;
 import java.util.Arrays;
 import java.util.List;
 
-import static com.azure.android.communication.common.CommunicationCloudEnvironmentModel.PUBLIC;
+import static com.azure.android.communication.chat.models.CommunicationCloudEnvironmentModel.PUBLIC;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsInstanceOf.instanceOf;
