+++ sdk/communication/azure-communication-chat/src/test/java/com/azure/android/communication/chat/TeamsUserIdentifierSerializationTests.java
@@ -1,7 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.communication.common;
+package com.azure.android.communication.chat;
+
+import com.azure.android.communication.chat.models.CommunicationCloudEnvironmentModel;
+import com.azure.android.communication.chat.models.CommunicationIdentifierModel;
+import com.azure.android.communication.chat.models.MicrosoftTeamsUserIdentifierModel;
+import com.azure.android.communication.common.CommunicationCloudEnvironment;
+import com.azure.android.communication.common.MicrosoftTeamsUserIdentifier;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
